Added Mute And Unmute functionality in ShortVideo Component

+++ src/Context/VideoContext.js
@@ -11,12 +11,14 @@ export const videoContext = createContext();
     const [NotificationMessage,setNotificationMessage] = useState('');
     const [isSubscribed,setisSubscribed] = useState(false);
     const [isLiked,setisLiked] = useState(false);
+    const [isSaved,setisSaved] = useState(false);
+    const [isDisLiked,setisDisLiked] = useState(false);
     const [Description,setDescription] = useState(false);
     const [shortvideoShowMessages,setshortvideoShowMessages] = useState(false);
-
+    const [Left,setLeft] = useState(null);
+    const [Top,setTop] = useState(null);
 
     const LikeVideo = async(user,videoId,video) => {
-      console.log(video)
       if(user){
         const docRef = doc(collection(firestore,`users/${user.uid}/LV`),videoId);
         const videoDocRef = doc(firestore,"videos",videoId);
@@ -26,7 +28,8 @@ export const videoContext = createContext();
         if(getdislikeDoc.data()){
           await deleteDoc(dislikedocRef);
             //  return "videoRemoved From dislikes"
-            setisLiked(true)
+            setisLiked(true);
+            setisDisLiked(false);
         }
         if(!getLikedDoc.exists()){
         const data = {videoURL:videoId}
@@ -51,11 +54,24 @@ export const videoContext = createContext();
         setshowToastNotification(true);
       }
     }
-    const checKLikedOrNot = async (user,videoId,video) => {
+    const checKLikedOrNot = async (user,videoId) => {
     if(user){
+      try{
       const docRef = doc(collection(firestore,`users/${user.uid}/LV`),videoId);
-      if((await getDoc(docRef)).exists()) setisLiked(true);
-      else setisLiked(false);
+      const dislikedocRef = doc(collection(firestore,`users/${user.uid}/DV`),videoId);
+      if((await getDoc(docRef)).exists()){
+        setisDisLiked(false);
+        setisLiked(true);
+      }
+      else if((await getDoc(dislikedocRef)).exists()){
+        setisLiked(false);
+        setisDisLiked(true);
+      } 
+      else if(!(await getDoc(dislikedocRef)).exists()) setisDisLiked(false);
+      else if(!(await getDoc(docRef)).exists()) setisLiked(false)
+      }catch(error){
+        console.log(error.message)
+      }
     }
     }
     const DisLikeVideo = async(user,videoId,video)=>{
@@ -66,6 +82,7 @@ export const videoContext = createContext();
       const docRef = doc(collection(firestore,`users/${user.uid}/LV`),videoId);
       const getLikedDoc = await getDoc(docRef);
      if(getLikedDoc.exists()){
+      if(video.likes > 0){
           await deleteDoc(doc(collection(firestore,`users/${user?.uid}/LV`),videoId));
           await updateDoc(videoDocRef,{
             likes: video.likes - 1,
@@ -74,16 +91,21 @@ export const videoContext = createContext();
           await setDoc(dislikedocRef,data);
           setNotificationMessage("video disliked");
           setshowToastNotification(true);
+          setisDisLiked(true);
+          setisLiked(false);
+        }
        }else{
         if(getdislikeDoc.exists()){
           await deleteDoc(dislikedocRef);
           setNotificationMessage("videoRemoved From dislikes");
           setshowToastNotification(true);
+          setisDisLiked(false);
         }else{
           const data = {videoURL:videoId}
           await setDoc(dislikedocRef,data);
           setNotificationMessage("video disliked");
           setshowToastNotification(true);
+          setisDisLiked(true);
         }
        }
       }else{
@@ -98,11 +120,11 @@ export const videoContext = createContext();
         if(!watchlaterDoc.exists()){
           await setDoc(watchlaterDocRef,{videoURL:videoId});
           setNotificationMessage("Saved to Watchlater");
-           return "videoAddedFromWatchLater"
+          setisSaved(true);
        }else{
         await deleteDoc(doc(collection(firestore,`users/${LoggedInUser?.uid}/WL`),videoId));
         setNotificationMessage("videoRemovedFromWatchLater");
-        return "videoRemovedFromWatchLater"
+        setisSaved(false);
        }
       }else{
         setNotificationMessage("You are not LoggedIn");
@@ -110,6 +132,21 @@ export const videoContext = createContext();
       document.body.style.opacity = "1";
       setshowToastNotification(true);
  }
+  const getWatchlaterVideo = async(LoggedInUser,searchQuery) => {
+         if(LoggedInUser){
+          try{
+           const LikedDocRef = doc(collection(firestore,`users/${LoggedInUser?.uid}/WL`),searchQuery);
+           const GetLikedDoc = await getDoc(LikedDocRef);
+           if(GetLikedDoc.exists()){
+             setisSaved(true);
+           }else{
+             setisSaved(false);
+           }
+         }catch(error){
+          console.log(error.message)
+         }
+             }
+       }
  const subscribeChannel = async(LoggedInUser,videoCreator) => {
   const docRef = doc(collection(firestore,`users/${LoggedInUser?.uid}/subscribedChannel`),videoCreator.uid);
   const channelDocRef = doc(collection(firestore,`users`),videoCreator.uid);
@@ -141,15 +178,43 @@ export const videoContext = createContext();
  const CheckSubscribedOrNot = async(LoggedInUser,SubscribedUser) => {
   if(LoggedInUser){
   if(SubscribedUser){
+    try{
   const docRef = doc(collection(firestore,`users/${LoggedInUser?.uid}/subscribedChannel`),SubscribedUser.uid);
   const Doc = await getDoc(docRef);
   if(Doc.exists()){
     setisSubscribed(true);
-  }else  setisSubscribed(false);;
+  }else  setisSubscribed(false);
+}catch(error){
+  console.log(error.message)
 }
 }
+}
+}
+const showmodal = (e) => {
+  if (window.innerWidth <= 600) {
+    document.body.style.opacity = "0.7";
+    setLeft(null);
+    setTop(null);
+  } else {
+    setLeft(e.pageX - 246);
+    setTop(e.clientY);
   }
+  document.body.style.overflow = "hidden";
+};
+const dots = document.getElementById("dots");
+if (dots) {
+  window.addEventListener("resize", () => {
+    if (window.innerWidth > 600) {
+      setLeft(dots.getBoundingClientRect().left - 246);
+      setTop(dots.getBoundingClientRect().y + 9);
+    } else {
+      setLeft(null);
+      setTop(null);
+    }
+  });
+}
  const getVideoPublishedTime = (FullLengthVideo) => {
+  try{
   if(FullLengthVideo.Videodata.Time){
      let seconds = Math.floor((Date.now() - FullLengthVideo.Videodata.Time)/1000);
       let value = seconds;
@@ -183,8 +248,12 @@ export const videoContext = createContext();
         else return "now";
       }
     }
+  }catch(error){
+    console.log(error.message);
+  }
 }
 const returnvideoTime = (duration) => {
+  try{
   const seconds = Math.floor(duration % 60);
   const minutes = Math.floor(duration / 60) % 60;
   const hours = Math.floor(duration / 3600)
@@ -193,7 +262,10 @@ const returnvideoTime = (duration) => {
   }else{
    return  minutes + ":" + seconds.toString().padStart(2, 0)
   }
+}catch(error){
+  console.log(error.message)
 }
-    return <videoContext.Provider value={{showModal,showToastNotification,shortvideoShowMessages,NotificationMessage,Description,bottomlayout,isSubscribed,isLiked,setisLiked,setisSubscribed,setDescription,setshortvideoShowMessages,setbottomlayout,setshowModal,setNotificationMessage,setshowToastNotification,LikeVideo,WatchLater,getVideoPublishedTime,returnvideoTime,subscribeChannel,DisLikeVideo,CheckSubscribedOrNot,checKLikedOrNot}}>{children}</videoContext.Provider>
+}
+    return <videoContext.Provider value={{showModal,showToastNotification,shortvideoShowMessages,NotificationMessage,Description,bottomlayout,isSubscribed,isLiked,isDisLiked,isSaved,Left,Top,setisLiked,setisSubscribed,setDescription,setshortvideoShowMessages,setbottomlayout,setshowModal,setNotificationMessage,setshowToastNotification,LikeVideo,WatchLater,getVideoPublishedTime,returnvideoTime,subscribeChannel,DisLikeVideo,CheckSubscribedOrNot,checKLikedOrNot,setisDisLiked,setisSaved,getWatchlaterVideo,showmodal}}>{children}</videoContext.Provider>
 }
 export default VideoActionProvider
